Move @RegisterBeanMapper to customizers package with the other Register* annotations
