+++ sqlobject/src/test/java/org/jdbi/v3/TestColumnMappers.java
@@ -22,7 +22,7 @@ import java.util.List;
 import org.jdbi.v3.sqlobject.SqlQuery;
 import org.jdbi.v3.sqlobject.customizers.RegisterColumnMapper;
 import org.jdbi.v3.sqlobject.customizers.RegisterColumnMapperFactory;
-import org.jdbi.v3.sqlobject.helpers.RegisterBeanMapper;
+import org.jdbi.v3.sqlobject.customizers.RegisterBeanMapper;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Rule;
