+++ sqlobject/src/test/java/org/jdbi/v3/docs/TestFoldToObjectGraph.java
@@ -29,7 +29,7 @@ import org.jdbi.v3.H2DatabaseRule;
 import org.jdbi.v3.Handle;
 import org.jdbi.v3.sqlobject.SqlObjectBuilder;
 import org.jdbi.v3.sqlobject.SqlQuery;
-import org.jdbi.v3.sqlobject.helpers.RegisterBeanMapper;
+import org.jdbi.v3.sqlobject.customizers.RegisterBeanMapper;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
